Add `Circos.initialize_from_matrix()` test code
